refactor(components): refactor avatar (#3364)
